pyproject.toml
[tool.poetry]
name = "optyfi_task"
version = "0.1.0"
description = ""
authors = ["Lukas Bures <burylukas@seznam.cz>"]

[tool.poetry.dependencies]
python = "^3.9"
backtrader = { git = "https://github.com/mementum/backtrader.git", rev = "master" }
pandas = "^1.4.3"
starlette = "^0.20.4"
cachetools = "^5.2.0"
matplotlib = "^3.5.3"
backtrader-plotting = "^2.0.0"
QuantStats = "^0.0.59"
pyfolio = "^0.9.2"
gql = "^3.4.0"
requests-toolbelt = "^0.9.1"
web3 = "^5.30.0"
python-binance = "^1.0.16"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
